[FIX] account_edi_ubl_cii: change for XRechnung version 3.0.1
Versions
--------
- 16.0+

Steps
-----
1. Create an XRechnung for a customer with no VAT and no email provided;
2. Validate the resulting XML against a validator (https://erechnungsvalidator.service-bw.de/)
3. Validation fails with "Buyer electronic address MUST be provided"

Issue
-----
See #167219.

Cause
-----
Recently there were problems with the validation of XRechnung invoices created
via odoo 16 (#160644 and #167219). With a change of the standard "XRechnung" to
version 3.0.1 (https://www.e-rechnung-bund.de/standard-xrechnung-3-0-1/) an
EndpointID is required (either the VAT number or an electronic address,
i.e. the customer email address).

Solution
--------
The application checks whether the customer has a VAT or email. If this is the case, it will be applied.
If not, the EnpointID "EM" should be used with the customer's email address.

Closes #167219

Part-of: odoo/odoo#170850
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Co-authored-by: Michael Holthausen <mholthausen@terrestris.de>